Rewrite in Typescript
