+++ .prettierrc.json
@@ -1,1 +1,4 @@
-{}
+{
+    "tabWidth": 4,
+    "singleQuote": false
+}
