+++ package.json
@@ -1,47 +1,58 @@
 {
-  "name": "passport-appointment-bot",
-  "version": "3.3.0",
-  "description": "",
-  "author": "Karl Kihlström",
-  "license": "MIT",
-  "main": "src/index.js",
-  "bin": "src/index.js",
-  "scripts": {
-    "start": "node ./src/index",
-    "dev": "nodemon ./src/index",
-    "dev:mock": "nodemon ./src/index --mock",
-    "package": "NODE_ENV=production pkg . && cp ./config.json ./dist",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "devDependencies": {
-    "nodemon": "^2.0.15",
-    "pkg": "^5.5.2",
-    "prettier": "^2.5.1"
-  },
-  "dependencies": {
-    "cheerio": "^1.0.0-rc.10",
-    "cors": "^2.8.5",
-    "express": "^4.17.3",
-    "fetch-cookie": "^2.0.3",
-    "node-fetch": "^2.6.7",
-    "puppeteer": "npm:puppeteer-core@^13.5.2",
-    "puppeteer-extra": "^3.2.3",
-    "puppeteer-extra-plugin-stealth": "^2.9.0",
-    "winston": "^3.6.0",
-    "yargs": "^17.4.0"
-  },
-  "pkg": {
-    "assets": [
-      "assets/**/*"
-    ],
-    "outputPath": "dist",
     "name": "passport-appointment-bot",
-    "targets": [
-      "node16-linux-arm64",
-      "node16-linux-x64",
-      "node16-macos-arm64",
-      "node16-macos-x64",
-      "node16-win-x64"
-    ]
-  }
+    "version": "3.3.0",
+    "description": "",
+    "author": "Karl Kihlström",
+    "license": "MIT",
+    "main": "src/index",
+    "bin": "src/index",
+    "scripts": {
+        "start": "ts-node src/index.ts",
+        "dev": "nodemon",
+        "dev:mock": "nodemon --mock",
+        "lint": "eslint -c .eslintrc.json --ext .ts ./src",
+        "package": "NODE_ENV=production pkg . && cp ./config.json ./dist",
+        "test": "echo \"Error: no test specified\" && exit 1"
+    },
+    "dependencies": {
+        "@lvksh/logger": "^1.6.3",
+        "chalk": "^4.1.2",
+        "cheerio": "^1.0.0-rc.10",
+        "cors": "^2.8.5",
+        "express": "^4.17.3",
+        "fetch-cookie": "^2.0.3",
+        "node-fetch": "^2.6.7",
+        "puppeteer": "npm:puppeteer-core@^13.5.2",
+        "puppeteer-extra": "^3.2.3",
+        "puppeteer-extra-plugin-stealth": "^2.9.0",
+        "winston": "^3.6.0",
+        "yargs": "^17.4.0"
+    },
+    "devDependencies": {
+        "@types/cors": "^2.8.12",
+        "@types/express": "^4.17.13",
+        "@types/node-fetch": "^2.6.1",
+        "@types/yargs": "^17.0.10",
+        "eslint": "^8.12.0",
+        "eslint-plugin-lvksh": "^1.3.0",
+        "nodemon": "^2.0.15",
+        "pkg": "^5.5.2",
+        "prettier": "^2.5.1",
+        "ts-node": "^10.7.0",
+        "typescript": "^4.6.3"
+    },
+    "pkg": {
+        "assets": [
+            "assets/**/*"
+        ],
+        "outputPath": "dist",
+        "name": "passport-appointment-bot",
+        "targets": [
+            "node16-linux-arm64",
+            "node16-linux-x64",
+            "node16-macos-arm64",
+            "node16-macos-x64",
+            "node16-win-x64"
+        ]
+    }
 }
